+++ api/rust/include/LIEF/rust/ELF/DynamicEntryFlags.hpp
@@ -20,7 +20,14 @@ class ELF_DynamicEntryFlags : public ELF_DynamicEntry {
   public:
   using lief_t = LIEF::ELF::DynamicEntryFlags;
 
+  auto flags() const {
+    return impl().raw_flags();
+  }
+
   static bool classof(const ELF_DynamicEntry& entry) {
     return lief_t::classof(&entry.get());
   }
+
+  private:
+  const lief_t& impl() const { return as<lief_t>(this); }
 };